small fixes and parallel condense

+++ rename.go
@@ -1,19 +1,41 @@
 package main
 
 import (
+	"flag"
 	"fmt"
 	"os"
 )
 
-func rename(ext string) error {
-	files, err := findFiles(ext)
-	if err != nil {
-		return err
+func renameCmd(args []string) error {
+	if len(args) < 3 {
+		return fmt.Errorf("expected a file extension")
 	}
 
-	for count, file := range files {
-		newName := fmt.Sprintf("%02d.%s", count+1, ext)
-		os.Rename(file, newName)
+	fs := flag.NewFlagSet("rename", flag.ExitOnError)
+
+	dryRun := fs.Bool("dry-run", false, "preview changes without modifying files")
+
+	fs.Parse(args)
+	extensions := fs.Args()
+
+	if len(extensions) < 1 {
+		return fmt.Errorf("expected directory or file(s)")
+	}
+
+	for _, extension := range extensions {
+		files, err := findFiles(extension)
+		if err != nil {
+			return err
+		}
+
+		for count, file := range files {
+			newName := fmt.Sprintf("%02d.%s", count+1, extension)
+			if *dryRun {
+				fmt.Println("mv", file, newName)
+			} else {
+				os.Rename(file, newName)
+			}
+		}
 	}
 
 	return nil
